+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Adds `Range.prototype.isEmpty`.
+- Adds `Range.prototype.isIntersecting`.
 
 ## 1.0.0 (May 11, 2015)
 - Adds support for setting range bounds with `new Range(1, 3, "[)")` to set